redis: reuse the same code for UI notifications
redis_signals.go has a Redis notifier via pub/sub.
redis_signal_outbound.go had an almost exact copy just for UI
notifications. Merge the second into the first with minimal changes, and
remove the small and pointless file.
